Refactor core functions to improve modularity and error handling
- Moved `init`, `pull`, and `push` logic into `core/sync.go` for better organization.
- Enhanced error handling for `sync` commands and updated CLI to log descriptive error messages.
- Modified `Clone` to detect and adapt to WSL environments.
- Updated `Run` to support additional arguments for command execution.
- Changed `loadDataConfig` to return an error for invalid configurations.
- Removed unused code and refactored repository logic for clarity.
