+++ internal/cli/sync/init.go
@@ -2,6 +2,7 @@ package sync
 
 import (
 	"github.com/mxlang/dotx/internal/core"
+	"github.com/mxlang/dotx/internal/logger"
 	"github.com/spf13/cobra"
 )
 
@@ -24,7 +25,9 @@ func newCmdInit(app core.App) *cobra.Command {
 		Args: cobra.ExactArgs(1),
 
 		Run: func(cmd *cobra.Command, args []string) {
-			app.Init(args[0], opts.deploy, opts.force)
+			if err := app.Init(args[0], opts.deploy, opts.force); err != nil {
+				logger.Error("failed to initialize dotfiles", "error", err)
+			}
 		},
 	}
