+++ internal/git/git.go
@@ -3,7 +3,6 @@ package git
 import (
 	"errors"
 
-	"github.com/go-git/go-git/v6/plumbing"
 	"github.com/mxlang/dotx/internal/fs"
 
 	"strings"
@@ -119,43 +118,4 @@ func Remote(repoDir fs.Path) ([]string, error) {
 	}
 
 	return remote.Config().URLs, nil
-}
-
-func IsBehindRemote(repoDir fs.Path) (bool, error) {
-	repo, err := git.PlainOpen(repoDir.AbsPath())
-	if err != nil {
-		return false, err
-	}
-
-	err = repo.Fetch(&git.FetchOptions{RemoteName: "origin", Progress: nil, Tags: git.NoTags})
-	if err != nil && !errors.Is(err, git.NoErrAlreadyUpToDate) {
-		return false, err
-	}
-
-	headRef, err := repo.Head()
-	if err != nil {
-		return false, err
-	}
-
-	remoteRef, err := repo.Reference(plumbing.NewRemoteReferenceName("origin", headRef.Name().Short()), true)
-	if err != nil {
-		return false, err
-	}
-
-	headCommit, err := repo.CommitObject(headRef.Hash())
-	if err != nil {
-		return false, err
-	}
-
-	remoteCommit, err := repo.CommitObject(remoteRef.Hash())
-	if err != nil {
-		return false, err
-	}
-
-	isAncestor, err := headCommit.IsAncestor(remoteCommit)
-	if err != nil {
-		return false, err
-	}
-
-	return isAncestor, nil
 }
